requirements.txt
Django==3.2.12
djangorestframework==3.13.1
django-environ==0.8.1
psycopg2-binary==2.9.3
django-mptt==0.13.4
drf-yasg==1.20.0